requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
authlib==1.2.1
python-dotenv==1.0.1
psycopg2-binary==2.9.10
requests==2.31.0
